Move S1 orbit stuff into sentinel1 subpackage
