+++ hyp3lib/depreciated/par_s1_slc_single.py
@@ -5,7 +5,7 @@ import os
 from hyp3lib import OrbitDownloadError
 from hyp3lib.depreciated.execute import execute
 from hyp3lib.depreciated.getParameter import getParameter
-from hyp3lib.orbits import download_orbit_file
+from hyp3lib.sentinel1.orbits import download_orbit_file
 
 
 def make_cmd(swath, acquisition_date, out_dir, pol=None):
